refactor: replace double with BigDecimal for monetary values

+++ src/main/java/ssrahoo/marketplaceapi/dto/UserRegistrationDto.java
@@ -1,8 +1,10 @@
 package ssrahoo.marketplaceapi.dto;
 
+import java.math.BigDecimal;
+
 public record UserRegistrationDto(
         String username,
         String email,
         String password,
-        Double wallet
+        BigDecimal wallet
 ) {}
